# src/main/resources/static/swagger.json
swagger: '2.0'
info:
  description: 'this is apis for heat-stone '
  version: 1.0.0
  title: Swagger for HeatStone
  contact:
    email: luxun_helloworld@163.com
host: heatstone-party.com
basePath: /v2
tags:
  - name: card
    description: all methods for card
schemes:
  - http
paths:
  /cards/:
    post:
      tags:
        - card
      summary: insert card
      description: insert new card entity
      operationId: insert
      produces:
        - application/json
      parameters:
        - in: body
          name: card
          description: 卡牌
          required: true
          schema:
            $ref: '#/definitions/Card'
      responses:
        200:
          schema:
            type: object
            additionalProperties:
              type: object
  /cards/{id}:
    get:
      tags:
      - "card"
      summary: "根据主键查询卡牌详情"
      description: "根据主键查询活动详情"
      operationId: "selectById"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        description: "卡牌id"
        required: true
        type: integer
      responses:
        200:
          schema:
            type: object
            additionalProperties:
              type: object
definitions:
  Card:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: 主键id
      name:
        type: string
        description: 名称
      costs:
        type: integer
        format: int32
        description: 卡片费用
      quality:
        type: integer
        format: int32
        description: 卡牌品质
      characters:
        type: integer
        format: int32
        description: 特效
      race:
        type: integer
        format: int32
        description: 种族
      img_url:
        type: string
        description: 图片链接
      bar_img_url:
        type: string
        description: 条形图片链接
      occupation_id:
        type: integer
        format: int32
        description: 所属职业
      dust:
        type: integer
        format: int32
        description: 粉尘
      source:
        type: integer
        format: int32
        description: 来源
      effect_desc:
        type: string
        description: 效果描述
      adapted:
        type: integer
        format: int32
        description: 适用环境
      create_time:
        type: string
        format: date-time
        description: 添加时间